Some refactoring. Color support in TradfriBulb. Added ParasollDoorWindowSensor.

+++ src/main/scala/interfaces/Battery.scala
@@ -4,5 +4,6 @@ package interfaces
 import monix.reactive.Observable
 
 trait Battery {
-  val battery: Observable[Int]
+  /** State of the battery, in the range 0...1 (1 means full) */
+  val battery: Observable[Double]
 }
